requirements.txt
Django==1.9.1
dominate==2.1.16
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
flask-paginate==0.4.1
Flask-PyMongo==0.4.0
Flask-Script==2.0.5
Flask-WTF==0.12
image==1.4.4
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.1.0
pymongo==3.2
visitor==0.1.2
Werkzeug==0.11.3
wheel==0.24.0
WTForms==2.1
